# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MeResponse {
  isAdmin: boolean;
}

export interface Company {
  id: number;
  name: string;
  careersUrl: string | null;
  createdAt: string;
}

export interface CreateCompanyInput {
  name: string;
  careersUrl?: string | null;
}

export interface UpdateCompanyInput {
  name?: string;
  careersUrl?: string | null;
}

export type UpdateJobStatusInputStatus =
  (typeof UpdateJobStatusInputStatus)[keyof typeof UpdateJobStatusInputStatus];

export const UpdateJobStatusInputStatus = {
  new: "new",
  saved: "saved",
  applied: "applied",
  removed: "removed",
} as const;

export interface UpdateJobStatusInput {
  status: UpdateJobStatusInputStatus;
}

export type KeywordType = (typeof KeywordType)[keyof typeof KeywordType];

export const KeywordType = {
  include: "include",
  exclude: "exclude",
} as const;

export interface Keyword {
  id: number;
  term: string;
  type: KeywordType;
  createdAt: string;
}

export type CreateKeywordInputType =
  (typeof CreateKeywordInputType)[keyof typeof CreateKeywordInputType];

export const CreateKeywordInputType = {
  include: "include",
  exclude: "exclude",
} as const;

export interface CreateKeywordInput {
  term: string;
  type?: CreateKeywordInputType;
}

export type SearchStatus = (typeof SearchStatus)[keyof typeof SearchStatus];

export const SearchStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Search {
  id: number;
  keywords: string;
  status: SearchStatus;
  jobCount: number;
  createdAt: string;
}

export type RunSearchInputLocation = {
  usRemote: boolean;
  usLocal: boolean;
  city: string;
  radiusMiles: number;
} | null;

export interface RunSearchInput {
  keywords: string;
  companyIds?: number[] | null;
  location?: RunSearchInputLocation;
}

export type JobListingStatus =
  (typeof JobListingStatus)[keyof typeof JobListingStatus];

export const JobListingStatus = {
  new: "new",
  saved: "saved",
  applied: "applied",
  removed: "removed",
} as const;

export interface JobListing {
  id: number;
  searchId: number;
  companyId: number;
  companyName: string;
  title: string;
  location: string | null;
  url: string;
  description: string | null;
  postedAt: string | null;
  matchedKeywords: string | null;
  status: JobListingStatus;
  createdAt: string;
}

export type SearchWithJobsStatus =
  (typeof SearchWithJobsStatus)[keyof typeof SearchWithJobsStatus];

export const SearchWithJobsStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface SearchWithJobs {
  id: number;
  keywords: string;
  status: SearchWithJobsStatus;
  jobCount: number;
  createdAt: string;
  jobs: JobListing[];
}