# lib/api-zod/src/generated/types/createKeywordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateKeywordInputType } from "./createKeywordInputType";

export interface CreateKeywordInput {
  term: string;
  type?: CreateKeywordInputType;
}