# lib/api-zod/src/generated/types/jobListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobListingStatus } from "./jobListingStatus";

export interface JobListing {
  id: number;
  searchId: number;
  companyId: number;
  companyName: string;
  title: string;
  location: string | null;
  url: string;
  description: string | null;
  postedAt: string | null;
  matchedKeywords: string | null;
  status: JobListingStatus;
  createdAt: Date;
}