# lib/api-zod/src/generated/types/runSearchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunSearchInputLocation } from "./runSearchInputLocation";

export interface RunSearchInput {
  keywords: string;
  companyIds?: number[] | null;
  location?: RunSearchInputLocation;
}